Report the time of a video's first displayed frame for LCP
This CL plumbs the information needed to capture LCP from videos
into Video timing information for LCP into the HTMLVideoElement, from
where it is sent to the PaintTimingDetector when the initial frame is
displayed.

Video byte size is taken from the WebMediaPlayer's PipelineStatistics,
while the result of the Timing-Allow-Origin check goes through URLIndex,
in the same manner as CORS checks.

Change-Id: I9e8ed63a2913d5dc70b0f1b2bf9f928db4f1b908