Initialize cache_state at navigation timing instantiation
Currently the transfer size of navigation timing entries is not being properly set based on the entry's cache state, always being set as if the resource was fetched from the network, even when cached. This is a result of the fact that the transferSize depends on PerformanceResourceTiming::cache_state_, which is not initialized for navigation timing, and hence always takes the default value. This CL fixes that by passing in the right cache state value when navigation timing entries are created.

Bug: 1262896
Change-Id: I290d97254bed846331549a3ff96c7c3e6502230d